refactor - 권한 인증 방식 id->username

+++ src/main/java/com/project/tableforyou/service/UserService.java
@@ -68,25 +68,34 @@ public class UserService {
 
     /* 회원 업데이트 */
     @Transactional
-    public void update(Long id, UserDto.UpdateRequest dto) {
+    public void update(String username, UserDto.UpdateRequest dto) {
 
-        log.info("Updating user with ID: {}", id);
-        User user = userRepository.findById(id).orElseThrow(() ->
-                new IllegalArgumentException("해당 회원이 존재하지 않습니다. id: " + id));
+        log.info("Updating user with username: {}", username);
+        User user = userRepository.findByUsername(username).orElseThrow(() ->
+                new IllegalArgumentException("해당 회원이 존재하지 않습니다. username: " + username));
         dto.setPassword(bCryptPasswordEncoder.encode(dto.getPassword()));
-        user.update(dto.getNickname(), dto.getPassword(), dto.getEmail());
-        log.info("User updated successfully with ID: {}", id);
+        if(!verifyAuthenticationByUsername(username, user.getUsername())) {
+            throw new RuntimeException("권한이 없습니다.");
+        } else {
+            user.update(dto.getNickname(), dto.getPassword(), dto.getEmail());
+            log.info("User updated successfully with username: {}", username);
+        }
     }
 
     /* 회원 삭제 */
     @Transactional
-    public void delete(Long id) {
+    public void delete(String username, Long id) {
 
         log.info("Deleting user with ID: {}", id);
         User user = userRepository.findById(id).orElseThrow(() ->
                 new IllegalArgumentException("해당 회원이 존재하지 않습니다. id: " + id));
-        userRepository.delete(user);
-        log.info("User deleted successfully with ID: {}", id);
+
+        if(!verifyAuthenticationByUsername(username, user.getUsername())) {
+            throw new RuntimeException("권한이 없습니다.");
+        } else {
+            userRepository.delete(user);
+            log.info("User deleted successfully with ID: {}", id);
+        }
     }
 
     /* 회원가입 오류 확인 */
@@ -111,5 +120,10 @@ public class UserService {
 
         log.info("Checking if user exists by nickname: {}", nickname);
         return userRepository.existsByNickname(nickname);
+    }
+
+    /* 자신의 권한인지 확인 */
+    private boolean verifyAuthenticationByUsername(String expectedUsername, String actualUsername) {
+        return actualUsername.equals(expectedUsername);
     }
 }
